Add support for Metatrader4 csv format.
Whilst it is a simple variation on the GenericCSVData, it is very
useful to have as a command-line option in btrun script. There are
many brokers that use Metatrader platforms for trading, and most
brokers allow exporting CSV files from the history centre with
historical quotes as provided by the broker.