Update plot to enable configure figure size
Update the plot function to enable it to have a confugrable figure size. The cerebro plot function had the params for width and height but were unused.

A lot of users are using the rcparams as the workaround to change the default figure size to make it work. This change will fix that issue.