In 2017 the numeric check was dropped from the code. The function for checking if the headers was numeric was still in the code but never used after the check. This is colsnumeric.
I have reinstated its used. I have added back in similar fashion (but not exactly the same) code that will set the self._colmapping dict which gets used when iterating through the pandas lines to create the backtrader lines.

The order of the columns is obviously important since there is no text to determine the lines. I have checked the historical codes as well as the docs and set the columns accordingly:

{'datetime': None, 'open': 0, 'high': 1, 'low': 2, 'close': 3, 'volume': 4, 'openinterest': 5}

The self.colsnumeric bool qualifier is carried forward and used in the Transform names section in start(). This will ensure that numeric headers skip the .lower() error and use dthe second conditional in

```
# Transform names (valid for .ix) into indices (good for .iloc)
        if self.p.nocase and not self.colsnumeric:
            colnames = [x.lower() for x in self.p.dataname.columns.values]
        else:
            colnames = [x for x in self.p.dataname
```

I will create some test cases but this is working so far.